adduser script
this script adds new users to the system and allows you to add cusotm names and uses a preconfigerd github backup zip to allow new users to be added.    this scri[t has not been tested outside a virtual machine so be carefull before using.   you must be root to do this